extract superpose3d and type modules to a common package
