+++ rdock-utils/rdock_utils/sdrmsd/sdrmsd.py
@@ -1,13 +1,28 @@
 import functools
 import logging
 import math
+from dataclasses import dataclass, field
 
 from openbabel import pybel
 
-from .superpose3d import MolAlignmentData, Superpose3D
-from .types import AutomorphismRMSD, CoordsArray, PoseMatchData, SDRMSDData
+from rdock_utils.common import AutomorphismRMSD, CoordsArray, MolAlignmentData, Superpose3D
 
 logger = logging.getLogger("SDRMSD")
+
+
+@dataclass
+class SDRMSDData:
+    skipped: list[int] = field(default_factory=list)
+    molecules_dict: dict[int, pybel.Molecule] = field(default_factory=dict)  # Save all poses with their dockid
+    population: dict[int, int] = field(default_factory=dict)  # Poses to be written
+    out_dict: dict[int, tuple[pybel.Molecule, float]] = field(default_factory=dict)
+
+
+@dataclass
+class PoseMatchData:
+    pose_index: int
+    docked_pose: pybel.Molecule
+    sdrmsd_data: SDRMSDData
 
 
 class SDRMSD:
